load custom flat image; complete armasters refactor
