+++ pypit/fluxspec.py
@@ -18,7 +18,7 @@ from pypit import msgs
 from pypit import ardebug as debugger
 from pypit.core import arflux
 from pypit import arload
-from pypit import armasters
+from pypit.core import armasters
 from pypit.core import arsave
 from pypit import arutils
 from pypit import masterframe
